Update 50.v61.mc.csv
Hi, not sure about this change: Tried to set the daytemp of the second heat circuit with the commands in tempsetpoints. However the v61 wrote the wrong info into the registers. 

e.g. i wrote 
w -c mc Setpoints.Monday 0.5;20;20
 .. the result was: 0.5;0.5;0.5 instead of 0.5;20;20

I traced the communication of 470 and V61. If I set the daytemp for HC2 at the 470 manually .. it sends 10 50b50502422a / 00 to the v61. then all days are set to x.x;21.0;21.0

The command ”B505","42",,,temp1 works fine for writing ...